COMP: Remove warning about argument not being int

+++ Utilities/cmtar/util.c
@@ -163,7 +163,7 @@ oct_to_int(char *oct)
 void
 int_to_oct_nonull(int num, char *oct, size_t octlen)
 {
-  snprintf(oct, octlen, "%*lo", octlen - 1, (unsigned long)num);
+  snprintf(oct, octlen, "%*lo", (int)(octlen-1), (unsigned long)num);
   oct[octlen - 1] = ' ';
 }
